move error-related logic and test to its own file
